+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Implemented proper orthogonal decomposition for evaluating total inelastic
+  scattering cross section
 
 ## [0.2.0] - 2022-01-01
 - Implemented estimator standard deviation